requirements.txt
binaryornot==0.4.4
cryptography==45.0.3
pycryptodome==3.23.0